Enhance capability mechanism to include Qwes and Modem features
- LocApiV02: handle feature status reporting.
- LCA changes to support capability mechanism.
- LCA test app : Capability mask is now uint64. So change format
specifiers for printing the uint64 mask value.

CRs-Fixed: 2807083

+++ location_api_msg_proto/src/LocationApiPbMsgConv.h
@@ -480,11 +480,11 @@ public:
     // LocationCallbacksMask to PBLocationCallbacksMask
     uint32_t getPBMaskForLocationCallbacksMask(const uint32_t &locCbMask) const;
     // LocationCapabilitiesMask to PBLocationCapabilitiesMask
-    uint32_t getPBMaskForLocationCapabilitiesMask(const uint32_t &locCapabMask) const;
+    uint64_t getPBMaskForLocationCapabilitiesMask(const uint64_t &locCapabMask) const;
 
     // **** helper function for mask conversion from protobuf masks to normal rigid values
     // PBLocationCapabilitiesMask to LocationCapabilitiesMask
-    uint32_t getLocationCapabilitiesMaskFromPB(const uint32_t &pbLocCapabMask) const;
+    uint64_t getLocationCapabilitiesMaskFromPB(const uint64_t &pbLocCapabMask) const;
     // PBLocationCallbacksMask to LocationCallbacksMask
     uint32_t getLocationCallbacksMaskFromPB(const uint32_t &pbLocCbMask) const;
